refactor: split the daemon protocol's compat policy from its wire vocabulary
protocol.ts answered two questions on two different clocks: whether two
builds may talk (version range, build skew, home ownership) and what the
messages are (frames, request names, task serialization). Every new RPC
edited the second half and showed up in the blame of the first.

The five compat symbols move to protocol-compat.ts; protocol.ts re-exports
them, so all 53 `daemon/protocol` import sites are byte-identical. 482 -> 405.
